Suppress warning about variable shadowing in libs
Geant4's headers pollute the global namespace with single-letter-named
global variables which are then easily shadowed in the code pulled in
with ROOT's includes. This results in an avalanche of warnings about variable
shadowing. It's not our fault and we can't do anything
about it but so we suppress this warning.